run the tests
